Fixed irrelevant commits being included
