--- src/frontend/next/src/components/AboutFooter/index.tsx
@@ -1,3 +1,0 @@
-import AboutFooter from './AboutFooter;';
-
-export default AboutFooter;
