--- src/frontend/next/src/components/Banner/Banner.tsx
@@ -1,134 +1,0 @@
-import React, { useState, useEffect, FC } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import { Fab, Grid, Typography } from '@material-ui/core';
-import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
-
-const useStyles = makeStyles((theme) => ({
-  h1: {
-    position: 'absolute',
-    color: theme.palette.primary.contrastText,
-    fontFamily: 'Roboto',
-    fontWeight: 'bold',
-    fontSize: '12rem',
-    display: 'block',
-    top: theme.spacing(20),
-    left: theme.spacing(8),
-    [theme.breakpoints.between('xs', 'sm')]: {
-      fontSize: '4rem',
-      textAlign: 'left',
-      left: theme.spacing(4),
-      right: theme.spacing(4),
-      top: theme.spacing(14),
-    },
-    [theme.breakpoints.between('md', 'lg')]: {
-      fontSize: '8rem',
-    },
-    [theme.breakpoints.up('xl')]: {
-      fontSize: '12rem',
-    },
-  },
-  heroBanner: {
-    height: '100vh',
-  },
-  version: {
-    position: 'absolute',
-    opacity: 0.85,
-    bottom: theme.spacing(6),
-    left: theme.spacing(8),
-    fontSize: '1.75rem',
-    color: theme.palette.primary.contrastText,
-    [theme.breakpoints.between('xs', 'sm')]: {
-      left: theme.spacing(4),
-      right: theme.spacing(4),
-      fontSize: '1.75rem',
-    },
-    [theme.breakpoints.between('md', 'lg')]: {
-      fontSize: '2rem',
-    },
-    [theme.breakpoints.up('xl')]: {
-      fontSize: '4rem',
-    },
-    textDecorationLine: 'none',
-    '&:hover': {
-      textDecorationLine: 'underline',
-    },
-  },
-  icon: {
-    height: '5.6rem',
-    width: '5.6rem',
-    position: 'relative',
-    bottom: theme.spacing(20),
-    zIndex: 300,
-    margin: '0 auto',
-    [theme.breakpoints.between('xs', 'sm')]: {
-      left: '55%',
-      bottom: theme.spacing(18),
-    },
-  },
-  anchor: {
-    position: 'relative',
-    top: '1rem',
-  },
-}));
-
-interface Props {}
-
-interface GitInfo {
-  status: string;
-  info: {
-    version: string;
-    sha: string;
-    gitHubUrl: string;
-  };
-}
-
-const Banner: FC<Props> = () => {
-  const classes = useStyles();
-  const [gitInfo, setGitInfo] = useState<GitInfo | null>({});
-
-  useEffect(() => {
-    async function getGitData() {}
-    getGitData();
-  }, []);
-
-  return (
-    <>
-      <div className={classes.heroBanner}>
-        <BannerDynamicItems />
-
-        <Typography variant="h1" className={classes.h1}>
-          Telescope
-        </Typography>
-
-        <a
-          href={`${gitInfo.gitHubUrl}`}
-          title={`git commit ${gitInfo.sha}`}
-          className={classes.version}
-        >
-          v {gitInfo.version}
-        </a>
-      </div>
-      <Grid
-        container
-        spacing={0}
-        direction="row"
-        alignItems="center"
-        justify="center"
-        className={classes.container}
-      >
-        <Grid id="back-to-top-anchor-mobile" item xs={8}>
-          <div className={classes.icon}>
-            <ScrollAction>
-              <Fab color="secondary" aria-label="scroll-down">
-                <KeyboardArrowDownIcon fontSize="large" color="textPrimary" />
-              </Fab>
-            </ScrollAction>
-          </div>
-        </Grid>
-      </Grid>
-      <div className={classes.anchor} id="back-to-top-anchor" />
-    </>
-  );
-};
-
-export default Banner;
