--- src/frontend/next/src/components/Banner/index.tsx
@@ -1,3 +1,0 @@
-import Banner from './Banner';
-
-export default Banner;
