+++ src/frontend/next/src/components/Footer/index.ts
@@ -1,0 +1,3 @@
+import Footer from './Footer';
+
+export default Footer;
