--- src/frontend/next/src/components/LogoIcon/index.tsx
@@ -1,3 +1,0 @@
-import LogoIcon from './LogoIcon';
-
-export default LogoIcon;
